+++ src/main/java/org/mmarini/rl/agents/ConstantAgent.java
@@ -107,7 +107,7 @@ public class ConstantAgent implements Agent {
     }
 
     @Override
-    public JsonNode getJson() {
+    public JsonNode json() {
         return null;
     }